+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/choosers/NoneChooser.kt
@@ -1,10 +1,13 @@
 package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers
 
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.models.PostId
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.Chooser
 import kotlinx.serialization.Serializable
+import org.joda.time.DateTime
 
 @Serializable
 class NoneChooser : Chooser {
-    override fun triggerChoose(): Collection<Int> {
+    override fun triggerChoose(time: DateTime, exceptions: List<PostId>): Collection<Int> {
         return emptyList()
     }
 }