+++ src/main/kotlin/com/github/insanusmokrassar/AutoPostTelegramBot/plugins/publishers/PublishPost.kt
@@ -2,7 +2,7 @@ package com.github.insanusmokrassar.AutoPostTelegramBot.plugins.publishers
 
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.exceptions.NoRowFoundException
 import com.github.insanusmokrassar.AutoPostTelegramBot.base.database.tables.PostsTable
-import com.github.insanusmokrassar.AutoPostTelegramBot.plugins.choosers.Chooser
+import com.github.insanusmokrassar.AutoPostTelegramBot.base.plugins.abstractions.Chooser
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.NewDefaultCoroutineScope
 import com.github.insanusmokrassar.AutoPostTelegramBot.utils.commands.Command
 import com.github.insanusmokrassar.TelegramBotAPI.bot.RequestsExecutor
@@ -67,23 +67,12 @@ class PublishPost(
             } ?: 1
 
             while (choosen.size < count) {
-                val toAdd = chooser.triggerChoose().filter {
-                    !choosen.contains(it)
-                }.let {
-                    val futureSize = choosen.size + it.size
-                    val toAdd = if (futureSize > count) {
-                        futureSize - count
-                    } else {
-                        it.size
-                    }
-                    it.toList().subList(0, toAdd)
-                }
-                if (toAdd.isEmpty()) {
+                val fromChooser = chooser.triggerChoose(exceptions = choosen)
+                if (fromChooser.isEmpty()) {
                     break
+                } else {
+                    choosen.addAll(fromChooser)
                 }
-                choosen.addAll(
-                    toAdd
-                )
             }
         } catch (e: NumberFormatException) {
             println("Can't extract number of posts")