+++ CHANGELOG.md
@@ -320,3 +320,10 @@ available messages for post
 ### 1.2.4 Hotfix
 
 * Hotfix for `PostPublisher` behaviour - now it will remove posts without forwardable messages
+
+### 1.2.5 Better choosers
+
+* `Chooser` abstraction was replaced to the base package
+* Add abstraction `PostId` which actually is `Int`
+* Now all choosers can accept exceptions for skip choosing of some posts
+* Added new command `getAutoPublications`